fix bug：压缩长度错误
zlib compress failed, len:3072 status:-5

+++ client/capture.c
@@ -137,7 +137,7 @@ void* zsend(void* ptr)
     
     char* zbuf = (char*)MALLOC(char, MAX_PACKET_LEN);
     memset(zbuf,'\0',MAX_PACKET_LEN);
-    int snplen = compressBound(sndbuf->length);
+    unsigned long snplen = compressBound(sndbuf->length);
     int ret = compress2((Bytef *)zbuf, (uLongf *)&snplen, (Bytef *)sndbuf->buffer, sndbuf->length, 9);
     if (ret == Z_OK)
     {
